+++ highfive/src/main/java/highfive/commands/consumer/FullHashDumpWriter.java
@@ -40,7 +40,8 @@ public class FullHashDumpWriter implements HashConsumer {
   }
 
   @Override
-  public void consumeTable(String genericName, boolean hasOrderingErrors) throws InvalidHashFileException {
+  public void consumeTable(String genericName, boolean hasOrderingErrors, long rowCount)
+      throws InvalidHashFileException {
     // Nothing to do
   }