+++ highfive/src/main/java/highfive/commands/consumer/HashLogger.java
@@ -58,7 +58,9 @@ public class HashLogger implements HashConsumer {
   }
 
   @Override
-  public void consumeTable(String genericName, boolean hasOrderingErrors) throws InvalidHashFileException {
+  public void consumeTable(String genericName, boolean hasOrderingErrors, long rowCount)
+      throws InvalidHashFileException {
+    // Nothing to do
   }
 
   @Override