+++ highfive/src/main/java/highfive/commands/consumer/SteppedRangeHashDumpWriter.java
@@ -52,7 +52,8 @@ public class SteppedRangeHashDumpWriter implements HashConsumer {
   }
 
   @Override
-  public void consumeTable(String genericName, boolean hasOrderingErrors) throws InvalidHashFileException {
+  public void consumeTable(String genericName, boolean hasOrderingErrors, long rowCount)
+      throws InvalidHashFileException {
     // Nothing to do
   }